backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.2
django-cors-headers==4.5.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.10
materialdash==0.0.22
msgpack==1.1.0
pillow==11.0.0
PyJWT==2.9.0
pysqlite3==0.5.4
redis==5.1.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
